requirements.txt
python-dotenv
pandas
requests
pytest